Support a Cover image for 2D work
Unless I am missing something, it seems you are not allowed to have a cover image for 2D NFTs. My artwork is very detailed and your process for making a cover image for the listings really "crunches" and blurs my images. It actually makes my artwork look bad and undesirable. I'd love to make a custom cover image for my NFT listing. For example, I uploaded a 74MB piece, then KO resizes it into a 440px tall, 59KB image for the listing cover, thumbnail image...it lowers the quality so much. It is really a shame and lowers the chance of a good first impression when people are browsing the site. (no code changes were made, just a feature suggestion.)